+++ packages/harpstrata/src/harpstrata/Apparatus/Interaction/types/index.ts
@@ -4,12 +4,15 @@ export enum InteractionIds {
   Bend1 = 'BEND1',
   Bend2 = 'BEND2',
   Bend3 = 'BEND3',
-  Overblow1 = 'OVERBLOW1',
-  Overblow2 = 'OVERBLOW2',
+  BlowBend1 = 'BLOWBEND1',
+  BlowBend2 = 'BLOWBEND2',
+  OverDraw1 = 'OVERDRAW1',
+  OverBlow1 = 'OVERBLOW1',
 }
 
 export interface Interaction {
   readonly id: InteractionIds;
 }
 
-export type InteractionMask = ReadonlyArray<Interaction>
+export type InteractionRow = ReadonlyArray<Interaction | undefined>
+export type InteractionMatrix = ReadonlyArray<InteractionRow>