+++ packages/harpstrata/src/harpstrata/Apparatus/constants/COUNTRY_TUNED_APPARATUS.ts
@@ -1,15 +1,18 @@
-import { ApparatusIds, Apparatus, InteractionMask, HalfstepIndexMatrix } from '../types'
+import { ApparatusIds, Apparatus, InteractionMatrix, HalfstepIndexMatrix } from '../types'
 import {
-  OVERBLOW2,
-  OVERBLOW1,
   BLOW,
   DRAW,
   BEND1,
   BEND2,
   BEND3,
+  BLOWBEND1,
+  BLOWBEND2,
+  OVERDRAW1,
+  OVERBLOW1,
 } from '../Interaction/constants'
 
 const halfstepIndexMatrix: HalfstepIndexMatrix = [
+  //    1          2          3          4          5          6          7          8          9         10
   [ undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, 34        ],
   [ 3        , undefined, undefined, 15       , undefined, 22       , undefined, 27       , 30       , 35        ],
   [ 0        , 4        , 7        , 12       , 16       , 19       , 24       , 28       , 31       , 36        ],
@@ -19,12 +22,19 @@ const halfstepIndexMatrix: HalfstepIndexMatrix = [
   [ undefined, undefined, 8        , undefined, undefined, undefined, undefined, undefined, undefined, undefined ],
 ] as const
 
-const interactionMask: InteractionMask = [
-  OVERBLOW2, OVERBLOW1, BLOW, DRAW, BEND1, BEND2, BEND3
+const interactionMatrix: InteractionMatrix = [
+  //    1          2          3          4          5          6          7          8          9         10
+  [ undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined, BLOWBEND2 ],
+  [ OVERBLOW1, undefined, undefined, OVERBLOW1, undefined, OVERBLOW1, undefined, BLOWBEND1, BLOWBEND1, BLOWBEND1 ],
+  [ BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW      ],
+  [ DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW      ],
+  [ BEND1    , BEND1    , BEND1    , BEND1    , BEND1    , BEND1    , OVERDRAW1, undefined, OVERDRAW1, OVERDRAW1 ],
+  [ undefined, BEND2    , BEND2    , undefined, undefined, undefined, undefined, undefined, undefined, undefined ],
+  [ undefined, undefined, BEND3    , undefined, undefined, undefined, undefined, undefined, undefined, undefined ],
 ] as const
 
 export const COUNTRY_TUNED_APPARATUS: Apparatus = {
   id: ApparatusIds.CountryTuned,
   halfstepIndexMatrix,
-  interactionMask,
+  interactionMatrix,
 } as const